Dict relocation (#694)
* Dict relocation

* Linter refactoring

* Fixes for the integration tests

* Fixes for dict relocation

* Address comments

* Modify the loop for final temporary segments relocation

+++ pkg/hintrunner/utils/testutils.go
@@ -5,14 +5,14 @@ import (
 	"github.com/NethermindEth/cairo-vm-go/pkg/vm/memory"
 )
 
-func WriteTo(vm *VM.VirtualMachine, segment uint64, offset uint64, val memory.MemoryValue) {
+func WriteTo(vm *VM.VirtualMachine, segment int, offset uint64, val memory.MemoryValue) {
 	err := vm.Memory.Write(segment, offset, &val)
 	if err != nil {
 		panic(err)
 	}
 }
 
-func ReadFrom(vm *VM.VirtualMachine, segment uint64, offset uint64) memory.MemoryValue {
+func ReadFrom(vm *VM.VirtualMachine, segment int, offset uint64) memory.MemoryValue {
 	val, err := vm.Memory.Read(segment, offset)
 	if err != nil {
 		panic(err)
